Move session type into _types/session.py
