+++ tamr_client/_types/__init__.py
@@ -20,4 +20,5 @@ from tamr_client._types.instance import Instance
 from tamr_client._types.json import JsonDict
 from tamr_client._types.operation import Operation
 from tamr_client._types.project import MasteringProject, Project
+from tamr_client._types.session import Session
 from tamr_client._types.url import URL
