+++ tamr_client/_types/session.py
@@ -1,0 +1,3 @@
+import requests
+
+Session = requests.Session
